Move spatial factories into spatial package
