+++ coverage.rc
@@ -2,10 +2,12 @@
 source = udata
 branch = True
 omit =
-    udata/tests/*
     */tests/*
 
 [report]
+omit =
+    */tests/*
+
 # Regexes for lines to exclude from consideration
 exclude_lines =
     # Have to re-enable the standard pragma
