+++ udata/core/spatial/tests/test_api.py
@@ -5,11 +5,11 @@ from flask import url_for
 
 from udata.utils import get_by
 
-from udata.tests.factories import (
-    faker, VisibleDatasetFactory, SpatialCoverageFactory, GeoZoneFactory,
-    GeoLevelFactory
-)
+from udata.tests.factories import faker, VisibleDatasetFactory
+
 from udata.tests.api import APITestCase
+
+from .factories import SpatialCoverageFactory, GeoZoneFactory, GeoLevelFactory
 
 
 class SpatialApiTest(APITestCase):
