+++ udata/core/spatial/tests/test_models.py
@@ -2,7 +2,8 @@
 from __future__ import unicode_literals
 
 from udata.tests import TestCase
-from udata.tests.factories import GeoZoneFactory, GeoLevelFactory
+
+from .factories import GeoZoneFactory, GeoLevelFactory
 
 from ..models import SpatialCoverage
